requirements.txt
pytorch-lightning==1.5.0
tensorboardX
test-tube==0.7.5
scikit-learn==0.24.2
sacrebleu==2.0.0
sentencepiece
matplotlib
rouge_score==0.0.4
nlp==0.4.0
datasets==2.2.2
bert-score==0.3.13
